longrunning: expose backoff Max parameter
The Max parameter is the steady-state constant interval. We don't
think it would be useful to allow adjustment of the Initial or
Multiplier values, since they should be in effect for a relatively
short time, the initial ramp-up to the constant polling interval.

Add WaitWithInterval, which accepts a duration and uses it for Max.

Also, make the default backoff parameters more reasonable.

Change-Id: I16d5c0792524be3ad2b919c3ba43aad863d6f737
Reviewed-on: https://code-review.googlesource.com/11470
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>